gdrive: add open
Fixes #3408
Related #2865
Fixes #3897
